Move audiovisual background playback policy to opt-in demo settings
